refactor: shrink internal realm boundaries
