+++ .agents/memory-bank/activeContext.md
@@ -6,6 +6,7 @@ The current branch is `codex/realm-convergence`. Keep this file focused on the a
 ## Current Branch Projects
 These records stay active until the branch is ready to merge. Before merge, move completed project records into the matching history file and leave only the next focus, constraints, and history index here.
 
+- **Models/Realm Internal Boundary Shrink (2026-07-03, complete):** Closed the high-priority internal Realm boundary leaks without changing Realm schema or repository APIs. Removed `CalibreServer.realm(in:)` from `CalibreServerRealm.swift`; the remaining test coverage now opens per-server sidecar Realms through `serverScopedRealmProvider.configuration(for:)` instead of making Realm schema depend on `AppContainer`. Moved legacy `CalibreBookRealm.migrateReadPos(...)` orchestration out of the schema object: `LegacyReadPositionRealmMappers.swift` now only parses `readPosData` into detached `BookDeviceReadingPosition` values, while `DatabaseBootstrapper.migrateLegacyReadPosData()` owns constructing the `CalibreBook`, saving positions through `ReadingPositionRepositoryProtocol`, and clearing `readPosData`. Split PDF preference responsibilities to match the Readium pattern: `PDFPreferenceValue` and PDF enums now live in Domain, PDF/UI adapter behavior (`fillColor`, dark-mode helpers, reader-engine preference conversion) lives in `Views/PDFView/PDFPreferenceAdapter.swift`, and `PDFOptions` value mapping lives in `Models/Realm/PDFOptionsRealmMappers.swift`; `PDFOptionsRealm.swift` no longer imports UIKit. Moved Calibre last-read-position entry/CFI encode-decode helpers from `BookReadingPositionRealm.swift` into `Models/Calibre/CalibreReadingPositionMappers.swift`. Remaining lower-priority cleanup candidates: `Persistable` bridge removal, locale date display helpers on Realm objects, and search cache transient fields. Validation: Realm internal boundary scans passed for `AppContainer` / `ReadingPositionRepositoryProtocol`, `import UIKit`, and legacy method names; the broad PDF/payload scan now only reports the expected `PDFPreferenceValue` references in `PDFOptionsRealmMappers.swift`; focused `MockAppContainerFactoryTests`, `DatabaseBootstrapperTests`, `RealmReadingPositionRepositoryTests`, `ReaderPreferenceRepositoryTests`, `YabrPDFViewControllerTests`, `YabrReadiumReaderViewControllerTests`, `RealmDomainMappingTests`, and `BookMetadataSyncWorkerTests` passed (75 tests); standard iOS Simulator build passed with existing warnings.
 - **Outer Schema Helper Boundary Shrink (2026-07-03, complete):** Removed the remaining outer-layer direct uses of `CalibreLibraryRealm.PrimaryKey`, `CalibreBookRealm.PrimaryKey`, and `CalibreBookRealm.RatingDescription` from App/View/Calibre value-model paths. `CalibreLibrary.identity(...)`, `CalibreBook.identity(...)`, and `CalibreBook.ratingDescription(for:)` now own the pure value identity/display logic while the Realm schema helper methods delegate to those helpers to preserve existing primary-key string formats and rating text. `AppContainerProtocol.library(forServerUUID:libraryName:)` now indexes `calibreLibraries` through the pure library identity; `LibraryViewModel` resolves failed/deleted sync book titles through `BookRepositoryProtocol.getBook(library:bookId:)`; `LibraryInfoBookListViewModel` uses the pure rating display helper. Added `RealmDomainMappingTests.testRealmSchemaHelpersDelegateToDomainHelpers` to guard compatibility. Validation: outer schema helper scans passed with no matches in the targeted App/View/Calibre paths; focused `CalibreDataSplitTests`, `LibraryInfoBookListViewModelTests`, `LibraryViewModelTests`, and `RealmDomainMappingTests` passed (51 tests); standard iOS Simulator build passed; `git diff --check` passed.
 - **Repository Mapper Helper Boundary Cleanup (2026-07-03, complete):** Moved the remaining low-risk repository-local mapper helpers back into the Realm mapper boundary. `BookBookmarkRealm.toValue()` / `init(value:)` and `BookHighlightRealm.toValue()` / `init(value:)` now live in `Models/Realm/AnnotationRealmMappers.swift` instead of `AnnotationRepository.swift`. `PositionIdentity` plus the `BookDeviceReadingPosition.identity` and `BookDeviceReadingPositionRealm.identity` helpers now live in `Models/Realm/ReadingPositionRealmMappers.swift` instead of `ReadingPositionRepository.swift`; `ReadingPositionRepository` now relies on type inference around the same identity grouping/deduplication logic. No repository APIs, Realm schema fields, or annotation/reading-position behavior changed. Validation: repository helper static scans passed with no mapper helper definitions left in the repository files and `RealmDomainMappingTests` still exercising `BookBookmarkRealm(value:)`, `BookHighlightRealm(value:)`, and `toValue()`; focused `RealmDomainMappingTests`, `RealmAnnotationRepositoryTests`, `RealmReadingPositionRepositoryTests`, and `ReadingSessionManagerTests` passed (43 tests); standard iOS Simulator build passed; `git diff --check` passed.
 - **ReadingPositionRepository historyBook Boundary Shrink (2026-07-03, complete):** Removed `ReadingPositionRepositoryProtocol.historyBook(for:bookId:)` and the `RealmReadingPositionRepository` implementation that opened the main metadata Realm and constructed `CalibreBookRealm.PrimaryKey` from the reading-position repository. `ReadingPositionHistoryViewModel` now composes the history page through `container.bookRepository.getBook(library:bookId:)`, preserving the existing fallback to `bookManager.readingBook`; `debugPositions`, `positionHistory`, and session sidecar Realm behavior remain inside the reading-position repository. `MockReadingPositionRepository` no longer carries `historyBook` fields, and the Book Detail history test now asserts book lookup through `MockBookRepository`. Validation: `historyBook` scan passed with no matches; `ReadingPositionRepository.swift` no longer references `CalibreBookRealm` / `CalibreBookRealm.PrimaryKey`; focused `RealmReadingPositionRepositoryTests`, `BookDetailViewModelTests`, and `ReadingSessionManagerTests` passed (40 tests); standard iOS Simulator build passed; `git diff --check` passed.
