+++ YetAnotherEBookReader/Models/Realm/CalibreBookRealm.swift
@@ -62,58 +62,6 @@ class CalibreBookRealm: Object, ObjectKeyIdentifiable {
         return (try? JSONSerialization.jsonObject(with: userMetaData, options: []) as? [String:Any]) ?? [:]
     }
     
-    func migrateReadPos(library: CalibreLibrary, repository: ReadingPositionRepositoryProtocol) {
-        guard let readPosData = readPosData,
-              let readPosDict = try? JSONSerialization.jsonObject(with: readPosData, options: []) as? NSDictionary,
-              let deviceMapDict = readPosDict["deviceMap"] as? NSDictionary
-        else {
-            return
-        }
-        
-        let book = CalibreBook(id: idInLib, library: library)
-        
-        deviceMapDict.forEach { key, value in
-            guard let deviceName = key as? String,
-                  let deviceReadingPositionDict = value as? [String: Any],
-                  var readerName = deviceReadingPositionDict["readerName"] as? String else {
-                return
-            }
-            
-            // MARK: TEMPFIX for reader name changes
-            if readerName == "FolioReader" {
-                readerName = ReaderType.YabrEPUB.rawValue
-            }
-            if readerName == "YabrPDFView" {
-                readerName = ReaderType.YabrPDF.rawValue
-            }
-            
-            var deviceReadingPosition = BookDeviceReadingPosition(id: deviceName, readerName: readerName)
-            
-            deviceReadingPosition.lastReadPage = deviceReadingPositionDict["lastReadPage"] as? Int ?? 0
-            deviceReadingPosition.lastReadChapter = deviceReadingPositionDict["lastReadChapter"] as? String ?? ""
-            deviceReadingPosition.lastChapterProgress = deviceReadingPositionDict["lastChapterProgress"] as? Double ?? 0.0
-            deviceReadingPosition.lastProgress = deviceReadingPositionDict["lastProgress"] as? Double ?? 0.0
-            deviceReadingPosition.furthestReadPage = deviceReadingPositionDict["furthestReadPage"] as? Int ?? deviceReadingPosition.lastReadPage
-            deviceReadingPosition.furthestReadChapter = deviceReadingPositionDict["furthestReadChapter"] as? String ?? deviceReadingPosition.lastReadChapter
-            deviceReadingPosition.maxPage = deviceReadingPositionDict["maxPage"] as? Int ?? 1
-            if let cfi = deviceReadingPositionDict["cfi"] as? String {
-                deviceReadingPosition.cfi = cfi
-            }
-            deviceReadingPosition.epoch = deviceReadingPositionDict["epoch"] as? Double ?? 0.0
-            if let lastPosition = deviceReadingPositionDict["lastPosition"] as? [Int] {
-                deviceReadingPosition.lastPosition = lastPosition
-            }
-            
-            deviceReadingPosition.structuralStyle = deviceReadingPositionDict["structuralStyle"] as? Int ?? .zero
-            deviceReadingPosition.structuralRootPageNumber = deviceReadingPositionDict["structuralRootPageNumber"] as? Int ?? .zero
-            deviceReadingPosition.positionTrackingStyle = deviceReadingPositionDict["positionTrackingStyle"] as? Int ?? .zero
-            deviceReadingPosition.lastReadBook = deviceReadingPositionDict["lastReadBook"] as? String ?? .init()
-            deviceReadingPosition.lastBundleProgress = deviceReadingPositionDict["lastBundleProgress"] as? Double ?? .zero
-            
-            repository.savePosition(deviceReadingPosition, for: book)
-        }
-    }
-    
     func updatePrimaryKey() {
         guard let serverUUID = serverUUID, let libraryName = libraryName else { return }
         primaryKey = CalibreBookRealm.PrimaryKey(serverUUID: serverUUID, libraryName: libraryName, id: idInLib.description)
