+++ YetAnotherEBookReader/Models/Realm/CalibreServerRealm.swift
@@ -34,25 +34,3 @@ extension CalibreServer: Persistable {
         return self.makeRealmObject()
     }
 }
-
-extension CalibreServer {
-    /// Opens the per-server sidecar Realm using the AppContainer's
-    /// configured provider. Callers that hold an AppContainer
-    /// instance (i.e. anywhere outside the static `AppContainer.shared`
-    /// global) should prefer this method so that the
-    /// `ServerScopedRealmConfigurationProviding` that backs the
-    /// container is honored. The previous `CalibreServer.realmPerf`
-    /// extension routed every call through `AppContainer.shared` and
-    /// therefore leaked state across test containers that shared a
-    /// server UUID.
-    func realm(in container: AppContainer) -> Realm {
-        let config = container.serverScopedRealmProvider.configuration(for: self)
-        let key = "CalibreServerRealm-\(ObjectIdentifier(container))-\(config.fileURL?.path ?? config.inMemoryIdentifier ?? "default")"
-        if let cachedRealm = Thread.current.threadDictionary[key] as? Realm {
-            return cachedRealm
-        }
-        let realm = try! Realm(configuration: config)
-        Thread.current.threadDictionary[key] = realm
-        return realm
-    }
-}
