+++ YetAnotherEBookReader/Models/Realm/PDFOptionsRealm.swift
@@ -4,7 +4,6 @@
 //
 
 import Foundation
-import UIKit
 import RealmSwift
 
 extension PDFThemeMode: PersistableEnum {}
@@ -12,83 +11,6 @@ extension PDFAutoScaler: PersistableEnum {}
 extension PDFLayoutMode: PersistableEnum {}
 extension PDFReadDirection: PersistableEnum {}
 extension PDFScrollDirection: PersistableEnum {}
-
-struct PDFPreferenceValue: Equatable {
-    var themeMode = PDFThemeMode.serpia
-    var selectedAutoScaler = PDFAutoScaler.Width
-    var pageMode = PDFLayoutMode.Page
-    var readingDirection = PDFReadDirection.LtR_TtB
-    var scrollDirection = PDFScrollDirection.Vertical
-
-    var hMarginAutoScaler = 5.0
-    var vMarginAutoScaler = 5.0
-    var hMarginDetectStrength = 2.0
-    var vMarginDetectStrength = 2.0
-    var marginOffset = 0.0
-    var lastScale = 1.0
-    var rememberInPagePosition = true
-
-    var isDark: Bool {
-        themeMode == .dark
-    }
-
-    func isDark<T>(_ darkValue: T, _ lightValue: T) -> T {
-        isDark ? darkValue : lightValue
-    }
-
-    var fillColor: CGColor {
-        switch themeMode {
-        case .none:
-            return .init(gray: 0.0, alpha: 0.0)
-        case .serpia:
-            return CGColor(red: 0.98046875, green: 0.9375, blue: 0.84765625, alpha: 1.0)
-        case .forest:
-            return CGColor(
-                red: CGFloat(Int("BA", radix: 16) ?? 255) / 255.0,
-                green: CGFloat(Int("D5", radix: 16) ?? 255) / 255.0,
-                blue: CGFloat(Int("C1", radix: 16) ?? 255) / 255.0,
-                alpha: 1.0
-            )
-        case .dark:
-            return .init(gray: 0.0, alpha: 1.0)
-        }
-    }
-
-    func toReaderEnginePreferences() -> ReaderEnginePreferences {
-        ReaderEnginePreferences(
-            themeMode: {
-                switch themeMode {
-                case .serpia:
-                    return 1
-                case .dark:
-                    return 2
-                default:
-                    return 0
-                }
-            }(),
-            fontSizePercentage: 100.0,
-            fontFamily: "Original",
-            lineHeight: 1.2,
-            pageMargins: 1.0,
-            scroll: pageMode == .Scroll,
-            scrollDirection: scrollDirection == .Horizontal ? 1 : 0,
-            volumeKeyPaging: false
-        )
-    }
-
-    mutating func apply(_ preferences: ReaderEnginePreferences) {
-        switch preferences.themeMode {
-        case 1:
-            themeMode = .serpia
-        case 2:
-            themeMode = .dark
-        default:
-            themeMode = .none
-        }
-        pageMode = preferences.scroll ? .Scroll : .Page
-        scrollDirection = preferences.scrollDirection == 0 ? .Vertical : .Horizontal
-    }
-}
 
 class PDFOptions: Object, ObjectKeyIdentifiable {
     @Persisted(primaryKey: true) var _id: ObjectId
@@ -125,35 +47,4 @@ class PDFOptions: Object, ObjectKeyIdentifiable {
         self.rememberInPagePosition = other.rememberInPagePosition
     }
 
-    func toValue() -> PDFPreferenceValue {
-        PDFPreferenceValue(
-            themeMode: themeMode,
-            selectedAutoScaler: selectedAutoScaler,
-            pageMode: pageMode,
-            readingDirection: readingDirection,
-            scrollDirection: scrollDirection,
-            hMarginAutoScaler: hMarginAutoScaler,
-            vMarginAutoScaler: vMarginAutoScaler,
-            hMarginDetectStrength: hMarginDetectStrength,
-            vMarginDetectStrength: vMarginDetectStrength,
-            marginOffset: marginOffset,
-            lastScale: lastScale,
-            rememberInPagePosition: rememberInPagePosition
-        )
-    }
-
-    func apply(_ value: PDFPreferenceValue) {
-        themeMode = value.themeMode
-        selectedAutoScaler = value.selectedAutoScaler
-        pageMode = value.pageMode
-        readingDirection = value.readingDirection
-        scrollDirection = value.scrollDirection
-        hMarginAutoScaler = value.hMarginAutoScaler
-        vMarginAutoScaler = value.vMarginAutoScaler
-        hMarginDetectStrength = value.hMarginDetectStrength
-        vMarginDetectStrength = value.vMarginDetectStrength
-        marginOffset = value.marginOffset
-        lastScale = value.lastScale
-        rememberInPagePosition = value.rememberInPagePosition
-    }
 }
